requirements.txt
langchain==0.3.2
langchain-community==0.3.1
tiktoken==0.8.0
sentence-transformers==3.1.1
einops==0.8.0
qdrant-client==1.11.3
python-dotenv==1.0.1
pandas==2.2.3
openai==1.52.2
fastembed-gpu==0.4.1
fastembed==0.4.1
tokenizers==0.20.1
python-rapidjson==1.20
fastapi==0.115.0
uvicorn==0.31.0